fix structure
